requirements.txt
# Face Recognition Attendance System
# Environment: faceenv_cam (Python 3.11)

streamlit
numpy==1.26.4
torch==2.2.0
torchvision==0.17.0
facenet-pytorch
faiss-cpu
pillow
plotly
pandas
